[pytket-quantinuum] simplify quantinuum auth and remove unused API methods
This is motivated by a couple of major factors:
* Storing passwords unencrypted is bad
* `keyring` is a difficult package to use cross platform as it requires
platform specific backends
* JWT id tokens expire quickly

Previously we would store id and refresh tokens using keyring. In theory
if an in date refresh token exists that is used to retrive a new ID
token. A frequent user would only have to log in once.

When this proved unstable we added the ability to also store your
password in keyring, and just recommended against it.
When this also proved unstable we added hack upon hack and recommended
frequent clearing of credentials and retries.

This PR aims to simplify this by requiring a log in for each session.
Username/email can be stored in pytket config and so will not be
requested each time. Password once per session, and never stored
(after making the login API request). Refresh and ID tokens are only
stored in memory, and the refresh token is used for long running
sessions to update the ID token. Internal time stamps are used to
make (hopefully) minimal login requests.

The keyring dependency is totally removed.
NOTE: running remote tests will always require a login from STDIN