[quantinuum] Support multiple API handlers
Reworks the way the default class level API handler worked by replacing
it with a global instance that is used by default for instances of
QuantinuumBackend unless specified otherwise.

Similarly creates a default memory store for credentials for this global
instance.

Also adds a few tests to demonstrate how this behavior works across
multiple backend instances in the default or custom login scenario.