[pytket-quantinuum] Catch ValueError for uvloop
When using nest_asyncio in the context of alternative event loops such as `uvloop`, `nest_asyncio.apply()` will raise a ValueError rather than a RuntimeError. For example:

```
ValueError: Can't patch loop of type <class 'uvloop.Loop'>
```


This commit adds ValueError to the `except` block in pytket-quantinuum to ignore this error and continue.